added functionality to pass a data type to get_model
Calling Model(Number) instantiates a model in which all parameters are of the "Number" type. I added a named parameter to get_model called parameter_type (which defaults to "nothing") that can be set to "Number". I also put the call to Model inside an if statement that calls Model("Number") if param_type = "Number" and calls Model() if param_type="nothing" and throws an error otherwise